feat: switch capability & assessment scoring to 1-10 scale with 0.5 increments
- ModelProfile.Capabilities and AssessmentCategory.Score now float64 on 0-10 scale
- Routing engine (score.go) normalizes by /10.0; thresholds doubled
- Classifier requirements and builtin catalog moved to 1-10
- Assessment eval functions return 0-10 (0.5 step) scores; confidence formula updated
- UI sliders max=10 step=0.5, all score displays show /10
- Config validation accepts 0-10

+++ internal/console/api_assessment.go
@@ -248,7 +248,7 @@ func (s *Server) handleApplyAssessmentProposal(w http.ResponseWriter, r *http.Re
 			}
 			mp := cfg.ModelProfiles[key]
 			if mp.Capabilities == nil {
-				mp.Capabilities = map[string]int{}
+				mp.Capabilities = map[string]float64{}
 			}
 			applyAll := len(accepted) == 0
 			for k, v := range rec.ProposedProfile.Capabilities {
